pyproject.toml
[project]
name = "toygres"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.21.0",
    "prompt-toolkit>=3.0.52",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "questionary>=2.1.1",
    "rapidfuzz>=3.14.3",
    "rich>=14.3.3",
]
